requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
django-versatileimagefield==2.2
djangorestframework==3.14.0
drf-flex-fields==1.0.0
Markdown==3.4.1
Pillow==9.2.0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
